move venv launch function to utils
